Fix bugs

+++ src/controllers/UserController.java
@@ -20,11 +20,11 @@ public class UserController {
 		return this.userService.getById(id);
 	}
 	
-	public User getByCPF(int cpf) throws Exception {
+	public User getByCPF(long cpf) throws Exception {
 		return this.userService.getByCPF(cpf);
 	}
 	
-	public void post(String firstName, String lastName, int rg, int cpf, Date birthdate, int phoneNumber, String address, int idPlan) throws Exception {
+	public void post(String firstName, String lastName, long rg, long cpf, Date birthdate, long phoneNumber, String address, int idPlan) throws Exception {
 		this.userService.post(firstName, lastName, rg, cpf, birthdate, phoneNumber, address, idPlan);
 	}
